requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0   # JWT token implementation for DRF
dj-rest-auth[with_social]==7.0.1     # REST API authentication utils
drf_spectacular==0.28.0     # API documentation